fix: refactored test for sonar cloud
