+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/controller/gamification/RankingController.java
@@ -1,7 +1,7 @@
 package com.itachallenge.challenge.controller.gamification;
 
 import com.itachallenge.challenge.dto.gamification.RankingResponseDto;
-import com.itachallenge.gamification.service.PointsService;
+import com.itachallenge.gamification.service.UserScoreService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -11,9 +11,9 @@ import reactor.core.publisher.Flux;
 @RequestMapping("/itachallenge/api/v1/users/ranking")
 public class RankingController {
 
-    private final PointsService pointsService;
+    private final UserScoreService pointsService;
 
-    public RankingController(PointsService pointsService) {
+    public RankingController(UserScoreService pointsService) {
         this.pointsService = pointsService;
     }
