--- itachallenge-challenge/src/main/java/com/itachallenge/gamification/service/PointsService.java
@@ -1,11 +1,0 @@
-package com.itachallenge.gamification.service;
-
-
-import com.itachallenge.challenge.dto.gamification.RankingResponseDto;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-
-public interface PointsService {
-
-    Flux<RankingResponseDto> getRankingDescOrder();
-}
