+++ itachallenge-challenge/src/main/java/com/itachallenge/gamification/service/UserScoreService.java
@@ -1,6 +1,8 @@
 package com.itachallenge.gamification.service;
 
 import com.itachallenge.challenge.dto.gamification.PointsHistoryResponseDto;
+import com.itachallenge.challenge.dto.gamification.RankingResponseDto;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.util.UUID;
@@ -8,5 +10,5 @@ import java.util.UUID;
 public interface UserScoreService {
 
     Mono<PointsHistoryResponseDto> getUserPointsHistory(UUID userId);
+    Flux<RankingResponseDto> getRankingDescOrder();
 }
-
