+++ itachallenge-challenge/src/main/java/com/itachallenge/gamification/service/UserScoreServiceImpl.java
@@ -2,11 +2,18 @@ package com.itachallenge.gamification.service;
 
 import com.itachallenge.challenge.dto.gamification.PointHistoryEntryDto;
 import com.itachallenge.challenge.dto.gamification.PointsHistoryResponseDto;
+import com.itachallenge.challenge.dto.gamification.RankingResponseDto;
 import com.itachallenge.gamification.document.UserScoreDocument;
+import com.itachallenge.gamification.exception.ServiceException;
 import com.itachallenge.gamification.repository.UserScoreRepository;
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+
 
 import java.util.Comparator;
 import java.util.List;
@@ -18,6 +25,7 @@ import java.util.UUID;
 public class UserScoreServiceImpl implements UserScoreService {
 
     private final UserScoreRepository userScoreRepository;
+    private static final Logger log = LoggerFactory.getLogger(UserScoreServiceImpl.class);
 
     @Override
     public Mono<PointsHistoryResponseDto> getUserPointsHistory(UUID userId) {
@@ -47,5 +55,14 @@ public class UserScoreServiceImpl implements UserScoreService {
                 .totalPoints(totalPoints)
                 .history(history)
                 .build();
+    }
+
+    @Override
+    public Flux<RankingResponseDto> getRankingDescOrder() {
+        return userScoreRepository.findUsersRanking()
+                .onErrorResume(DataAccessException.class, ex -> {
+                    log.error("DB error fetching ranking: {}", ex.getMessage());
+                    return Flux.error(new ServiceException("Could not retrieve ranking"));
+                });
     }
 }
