+++ itachallenge-challenge/src/main/resources/application.yml
@@ -11,7 +11,7 @@ spring:
     mongodb:
       #uri: mongodb://admin_challenge:BYBcMJEEWw5egRUo@localhost:27017/challenges?authSource=admin
       uri: mongodb://test_challenge:SBM55CTYA8Ui32yC@87.106.229.175:27017/challenges?authSource=admin
-      # uri: mongodb://localhost:27017/challenge
+      #uri: mongodb://mongodb-test:27017/challenge
       uuid-representation: standard
   #profiles:
     #active: mongock
