+++ itachallenge-challenge/src/test/java/com/itachallenge/gamification/repository/UserScoreRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.itachallenge.gamification.repository;
 
+import com.itachallenge.challenge.dto.gamification.RankingResponseDto;
 import com.itachallenge.gamification.document.UserScoreDocument;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -69,6 +70,14 @@ class UserScoreRepositoryTest {
                 .createdAt(LocalDateTime.now())
                 .build();
 
+        when(userScoreRepository.saveAll(List.of(score1, score2)))
+                .thenReturn(Flux.just(score1, score2));
+
+        RankingResponseDto pepito = new RankingResponseDto("Pepito", 100);
+
+        when(userScoreRepository.findUsersRanking())
+                .thenReturn(Flux.just(pepito));
+
         userScoreRepository.saveAll(List.of(score1, score2)).blockLast();
 
         StepVerifier.create(userScoreRepository.findUsersRanking())
@@ -79,6 +88,9 @@ class UserScoreRepositoryTest {
     @Test
     @DisplayName("Should return empty ranking when no scores exist")
     void givenNoScores_whenFindUsersRanking_thenReturnsEmpty() {
+        when(userScoreRepository.findUsersRanking())
+                .thenReturn(Flux.empty());
+
         StepVerifier.create(userScoreRepository.findUsersRanking())
                 .expectNextCount(0)
                 .verifyComplete();
